backend/requirements.txt
# Django framework
Django>=4.2,<5.0

# Django Ninja for REST API
django-ninja>=1.5.3

# CORS headers for frontend integration
django-cors-headers>=4.9.0

# # Environment variables
# python-decouple==3.8

# Testing
pytest>=9.0.2
pytest-django>=4.12.0
